mklive: generate eltorito efi image manually + tweaks/fixes
It turns out that limine fails to come up in bios mode on some
machines if the EFI/BOOT directory in the root is fully populated,
which is strange but fix it by only including what we want.

While at it, do not use the limine-provided limine-uefi-cd.bin
as that contains all images too, it does not seem to harm anything
but generate one manually anyway, only containing the same efi
images as the root directory.

While at it, include hfsplus for x86_64 images, for efi on legacy
macs.