mklive.sh: drop partition_offset=16
When combined with the rest of our hybrid arrangement this results
in a mountable gap1 partition (sda1 or whatever) containing an
iso9660 filesystem but this partition is not useful as the actual
live erofs is not mountable through it (superblock read fails).

This results in stuff like usb stick getting mounted twice when
inserted into a booted OS, as well as potential boot failures
when autodetecting the root.